Move client packets in own namespace
