+++ Sphere/Sphere.Client/Services/Handlers/CharacterCreatePacketHandler.cs
@@ -1,5 +1,4 @@
-﻿using Godot;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Sphere.Common.Entities;
 using Sphere.Common.Enums;
 using Sphere.Common.Interfaces.Packets;
@@ -8,6 +7,7 @@ using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Models;
 using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 using Sphere.Services.Services.Readers;
 
 namespace Sphere.Services.Services.Handlers
