+++ Sphere/Sphere.Client/Services/Handlers/CharacterSelectPacketHandler.cs
@@ -5,11 +5,7 @@ using Sphere.Common.Interfaces.Repository;
 using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Packets;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Sphere.Common.Packets.Client;
 
 namespace Sphere.Services.Services.Handlers
 {
