+++ Sphere/Sphere.Client/Services/Handlers/IngamePingPacketHandler.cs
@@ -5,6 +5,7 @@ using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 
 namespace Sphere.Services.Services.Handlers
 {
