+++ Sphere/Sphere.Client/Services/PacketParser.cs
@@ -1,8 +1,7 @@
 ﻿using Sphere.Common.Enums;
 using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
-using Sphere.Common.Packets;
-using System.Reflection.Metadata.Ecma335;
+using Sphere.Common.Packets.Client;
 
 namespace Sphere.Client.Services
 {
@@ -28,7 +27,7 @@ namespace Sphere.Client.Services
                 0x15 => new CharacterSelectPacket(packet),
                 0x13 => new IngameAcknowledgePacket(packet),
 
-                //_ when packet.OriginalMessage[15] == 0x80 => new CharacterCreatePacket(packet),
+                _ when packet.OriginalMessage[15] == 0x80 => new CharacterCreatePacket(packet),
                 _ when packet.OriginalMessage[15] == 0x40 => new LoginPacket(packet),
                 _ => null// throw new ArgumentException($"Unknown packet: {packet}")
             };
