+++ Sphere/Sphere.Common/Interfaces/Packets/IPacket.cs
