+++ Sphere/Sphere.Common/Interfaces/Tcp/IClientAccessor.cs
@@ -1,12 +1,16 @@
-﻿using Godot;
-using Sphere.Common.Entities;
+﻿using Sphere.Common.Entities;
 using Sphere.Common.Enums;
+using Sphere.Common.Interfaces.Nodes;
 
 namespace Sphere.Common.Interfaces.Tcp
 {
     public interface IClientAccessor
     {
+        IServer Server { get; init; }
+
         ITcpClient Client { get; set; }
+
+        IClient GameClient { get; set; }
 
         ushort ClientId { get; set; }
