+++ Sphere/Sphere.Common/Packets/Client/ClientPingPacketLong.cs
@@ -2,7 +2,7 @@
 using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 
-namespace Sphere.Common.Packets
+namespace Sphere.Common.Packets.Client
 {
     public class ClientPingPacketLong : Packet, IPacket
     {
