+++ Sphere/Sphere.Common/Packets/Client/ClientPingPacketShort.cs
@@ -2,18 +2,18 @@
 using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 
-namespace Sphere.Common.Packets
+namespace Sphere.Common.Packets.Client
 {
-    public class IngamePingPacket : Packet, IPacket
+    public class ClientPingPacketShort : Packet
     {
         public readonly PacketBase BasePacket;
 
-        public IngamePingPacket(PacketBase basePacket) : base(basePacket)
+        public ClientPingPacketShort(PacketBase basePacket) : base(basePacket)
         {
             BasePacket = basePacket ?? throw new ArgumentNullException(nameof(basePacket));
         }
 
-        public override PacketType PacketType => PacketType.IngamePingPacket;
+        public override PacketType PacketType => PacketType.ClientPingShort;
 
         public override Task Handle(IPacketHandler handler, CancellationToken cancellationToken)
         {
