+++ Sphere/Sphere.Common/Packets/Client/IngamePingPacket.cs
@@ -2,18 +2,18 @@
 using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 
-namespace Sphere.Common.Packets
+namespace Sphere.Common.Packets.Client
 {
-    public class IngameAcknowledgePacket : Packet, IPacket
+    public class IngamePingPacket : Packet, IPacket
     {
         public readonly PacketBase BasePacket;
 
-        public IngameAcknowledgePacket(PacketBase basePacket) : base(basePacket)
+        public IngamePingPacket(PacketBase basePacket) : base(basePacket)
         {
             BasePacket = basePacket ?? throw new ArgumentNullException(nameof(basePacket));
         }
 
-        public override PacketType PacketType => PacketType.IngameAcknowledge;
+        public override PacketType PacketType => PacketType.IngamePingPacket;
 
         public override Task Handle(IPacketHandler handler, CancellationToken cancellationToken)
         {
