+++ Sphere/Sphere.Common/Packets/Client/LoginPacket.cs
@@ -2,11 +2,11 @@
 using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 
-namespace Sphere.Common.Packets
+namespace Sphere.Common.Packets.Client
 {
-    public class CharacterCreatePacket : Packet, IPacket
+    public class LoginPacket : Packet, IPacket
     {
-        public CharacterCreatePacket(PacketBase packet) : base(packet)
+        public LoginPacket(PacketBase basePacket) : base(basePacket)
         {
         }
 
@@ -18,16 +18,13 @@ namespace Sphere.Common.Packets
 
         public byte[] BeginPayload => _packet.OriginalMessage[13..15]; // 0x08 0x40
 
-        public override PacketType PacketType => PacketType.CreateCharacter;
+        public override PacketType PacketType => PacketType.Login;
 
         public byte LoginLength => _packet.OriginalMessage[16]; // 0x45
 
-        public byte CharacterIndex => _packet.OriginalMessage[17];
+        public byte StartLogin => _packet.OriginalMessage[17]; // 0xFC
 
-        public byte[] Nickname => _packet.OriginalMessage[20..^5];
-
-        public byte[] CharacterInfo => _packet.OriginalMessage[^5..];
-
+        public byte[] Payload => _packet.OriginalMessage[18..];
 
         public override Task Handle(IPacketHandler handler, CancellationToken cancellationToken)
         {
