+++ Sphere/Sphere.Test.Common/TestExtensions.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using Sphere.Common.Enums;
-using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 
 namespace Sphere.Test.Common
 {
