+++ Sphere/Sphere.Test.Unit/Handlers/ClientPingLongPacketHandlerTests.cs
@@ -4,6 +4,7 @@ using Sphere.Common.Enums;
 using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 using Sphere.Services.Services.Handlers;
 using Sphere.Test.Common;
