+++ Sphere/Sphere.Test.Unit/Handlers/ClientPingShortPacketHandlerTests.cs
@@ -8,6 +8,7 @@ using Sphere.Common.Interfaces.Repository;
 using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 using Sphere.Services.Services.Handlers;
 using Sphere.Test.Common;
 using static Godot.WebSocketPeer;
