+++ Sphere/Sphere.Test.Unit/Handlers/IngamePingHandlerTests.cs
@@ -4,6 +4,7 @@ using Sphere.Common.Interfaces.Packets;
 using Sphere.Common.Interfaces.Services;
 using Sphere.Common.Interfaces.Tcp;
 using Sphere.Common.Packets;
+using Sphere.Common.Packets.Client;
 using Sphere.Services.Services.Handlers;
 
 namespace Sphere.Test.Unit.Handlers
